+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to Rust's notion of
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html). All versions prior
 to 1.0.0 are beta releases.
 
+## [Unreleased]
+- MSRV has been increased to 1.65.0
+- Migrated to `nom 8`, `which 5`.
+
 ## [0.6.1] - 2026-01-03
 ### Added
 - Support for configuring Unix-specific pinentry options, behind `#[cfg(unix)]`: